Moved table component from Admin Page
