+++ FrontEnd/src/routes/AdminPage/AdminPage.jsx
@@ -3,8 +3,8 @@ import './AdminGlobal.css';
 import Header from '../../components/Admin/Header/Header';
 import Menu from './Menu/Menu';
 import UserDetail from './DetailView/UserDetail';
-import UserTable from './Table/UserTable';
-import ProfilesTable from './Table/ProfilesTable';
+import UserTable from '../../components/Admin/Table/UserTable';
+import ProfilesTable from '../../components/Admin/Table/ProfilesTable';
 import ProfileDetail from './DetailView/ProfileDetail';
 import css from './AdminPage.module.css';
 import { Routes, Route } from 'react-router-dom';
